Deleted folders and files that are not required in the published repo. Renamed examples to match paper phrasing.
