--- Test2025/Project.toml
@@ -1,3 +1,0 @@
-[deps]
-JuMP = "4076af6c-e467-56ae-b986-b466b2749572"
-PiecewiseLinearOpt = "0f51c51e-adfa-5141-8a04-d40246b8977c"
