+++ examples/CLSP-Bin/clsp_scenarios_synthetic.csv
