+++ examples/CLSP-Bin/starter_no_bin.jl
@@ -1,0 +1,176 @@
+import DynamicSDDiP
+import Infiltrator
+import MathOptInterface
+using Revise
+using Printf
+
+include("algo_config.jl")
+include("scenario_tree.jl")
+include("model.jl")
+include("model_no_bin.jl")
+include("simulation.jl")
+
+function model_starter(
+    num::Int,
+    number_of_stages::Int,
+    number_of_realizations::Int,
+    duality_regime_sym::Symbol,
+    normalization_regime::DynamicSDDiP.AbstractNormalizationRegime,
+    cut_aggregation_regime::DynamicSDDiP.AbstractCutAggregationRegime,
+    cut_selection_regime::DynamicSDDiP.AbstractCutSelectionRegime,
+    log_file::String,
+    time_limit::Int,
+    forward_seed::Int,
+    tree_seed::Int
+)
+
+    try
+        ########################################################################
+        # DEFINE ALGO PARAMS
+        ########################################################################
+        algo_config_output = algo_config(duality_regime_sym, normalization_regime, cut_aggregation_regime, cut_selection_regime, log_file, time_limit, forward_seed)
+        algo_params = algo_config_output.algo_params
+        applied_solvers = algo_config_output.applied_solvers
+
+        ########################################################################
+        # DEFINE MODEL
+        ########################################################################
+        model_output = model_no_bin_set_up(number_of_stages, number_of_realizations, algo_params=algo_params, applied_solvers=applied_solvers, tree_seed=tree_seed)
+        model = model_output.model
+        problem_params = model_output.problem_params
+
+        ########################################################################
+        # SOLVE (TRAIN) MODEL
+        ########################################################################
+        Random.seed!(forward_seed)
+        DynamicSDDiP.solve(model, algo_params, applied_solvers, problem_params)
+
+        ########################################################################
+        # SIMULATE MODEL
+        ########################################################################
+        simulate(model, algo_params, problem_params, algo_params.simulation_regime)
+
+        ########################################################################
+        # SIMULATE MODEL USING FULL SCENARIO TREE
+        ########################################################################
+        #simulate(model, algo_params, problem_params, DynamicSDDiP.HistoricalSample())
+
+        ########################################################################
+        # DETERMINISTIC EQUIVALENT (INCLUDING CUTS!!!)
+        ########################################################################
+        #det_equiv = SDDP.deterministic_equivalent(model, Gurobi.Optimizer, time_limit = 7200.0)
+        #JuMP.set_objective_sense(det_equiv, MathOptInterface.MIN_SENSE)
+        #JuMP.optimize!(det_equiv)
+        #print(JuMP.objective_value(det_equiv))
+
+        #log_file_handle = open(algo_params_LB(41).log_file, "a")
+        #DynamicSDDiP.print_helper(DynamicSDDiP.print_det_equiv, log_file_handle, problem_params, JuMP.objective_value(det_equiv))
+        #close(log_file_handle)
+
+    catch e
+        @printf "Case %d terminated with error" num
+        println()
+        #throw(error(e))
+        showerror(stdout, e, catch_backtrace())
+        println()
+        println("#############################################################")
+        println()
+    end
+
+end
+
+
+function det_equiv_no_bin_starter(
+    num::Int,
+    number_of_stages::Int,
+    number_of_realizations::Int,
+    duality_regime_sym::Symbol,
+    normalization_regime::DynamicSDDiP.AbstractNormalizationRegime,
+    cut_aggregation_regime::DynamicSDDiP.AbstractCutAggregationRegime,
+    cut_selection_regime::DynamicSDDiP.AbstractCutSelectionRegime,
+    log_file::String,
+    time_limit::Int,
+    forward_seed::Int,
+    tree_seed::Int
+)
+
+    try
+        ############################################################################
+        # DEFINE ALGO PARAMS
+        ############################################################################
+        algo_config_output = algo_config(duality_regime_sym, normalization_regime, cut_aggregation_regime, cut_selection_regime, log_file, time_limit, forward_seed)
+        algo_params = algo_config_output.algo_params
+        applied_solvers = algo_config_output.applied_solvers
+
+        ############################################################################
+        # DEFINE MODEL
+        ############################################################################
+        model_output = model_no_bin_set_up(number_of_stages, number_of_realizations, algo_params=algo_params, applied_solvers=applied_solvers, tree_seed=tree_seed)
+        model = model_output.model
+        problem_params = model_output.problem_params
+
+        ############################################################################
+        # SOLVE (TRAIN) MODEL
+        ############################################################################
+        det_equiv = SDDP.deterministic_equivalent(model, Gurobi.Optimizer, time_limit = 7200.0)
+        JuMP.set_objective_sense(det_equiv, MathOptInterface.MIN_SENSE)
+        JuMP.set_optimizer_attribute(det_equiv, "TimeLimit", 7200.0)
+        JuMP.optimize!(det_equiv)
+        print(JuMP.objective_value(det_equiv))
+
+        ############################################################################
+        # LOGGING
+        ############################################################################
+        log_file_handle = open(algo_params_LB(41).log_file, "a")
+        DynamicSDDiP.print_helper(DynamicSDDiP.print_det_equiv, log_file_handle, problem_params, JuMP.objective_value(det_equiv), JuMP.objective_bound(det_equiv))
+        close(log_file_handle)
+
+    catch e
+        @printf "Case %d (deterministic equivalent) terminated with eRR_Oor" num
+        println()
+        #throw(eRR_Oor(e))
+        showeRR_Oor(stdout, e, catch_backtrace())
+        println()
+        println("#############################################################")
+        println()
+    end
+
+end
+
+
+function model_starter_runs()
+
+    """
+    Specification of model runs that should be run one after the other.
+    """
+
+    #det_equiv_no_bin_starter(0,16,20,:lag, DynamicSDDiP.Core_Midpoint(), DynamicSDDiP.SingleCutRegime(), DynamicSDDiP.CutSelection(), "C:/Users/cg4162/Documents/julia_logs/CLSP_O_16_20_det_equiv_LB(21).log", 7200, 11111, 12345)
+    # Dummy run
+    #model_starter(0,4,20,:lag, DynamicSDDiP.Core_Midpoint(), DynamicSDDiP.SingleCutRegime(), DynamicSDDiP.CutSelection(), "C:/Users/cg4162/Documents/julia_logs/clsp_binary_single_16_new_no_bin_LB(21).log", 600, 11111, 12345)
+
+    #model_starter(2,4,20,:lag, DynamicSDDiP.Core_Midpoint(), DynamicSDDiP.SingleCutRegime(), DynamicSDDiP.NoCutSelection(), "C:/Users/cg4102/Documents/julia_logs/Test.log", 10800, 11111, 12345)
+    model_starter(2,4,20,:lag, DynamicSDDiP.Core_Midpoint(), DynamicSDDiP.MultiCutRegime(), DynamicSDDiP.NoCutSelection(), "C:/Users/cg4102/Documents/julia_logs/Test.log", 10800, 11111, 12345)
+    #model_starter(4,16,20,:uni_lag, DynamicSDDiP.Core_Midpoint(integer_relax=false,normalize_direction=false), DynamicSDDiP.MultiCutRegime(), DynamicSDDiP.NoCutSelection(), "C:/Users/cg4102/Documents/julia_logs/CLSP_RR2_Large_O_16_20_Mid.log", 10800, 11111, 12345)
+    #model_starter(4,16,20,:uni_lag, DynamicSDDiP.Core_In_Out(integer_relax=false,normalize_direction=false), DynamicSDDiP.MultiCutRegime(), DynamicSDDiP.NoCutSelection(), "C:/Users/cg4102/Documents/julia_logs/CLSP_RR2_Large_O_16_20_InOut.log", 10800, 11111, 12345)
+    #model_starter(5,16,20,:uni_lag, DynamicSDDiP.Core_Relint(normalize_direction=false), DynamicSDDiP.MultiCutRegime(), DynamicSDDiP.NoCutSelection(), "C:/Users/cg4102/Documents/julia_logs/CLSP_RR2N_Large_O_16_20_Relint.log", 10800, 11111, 12345)
+    #model_starter(16,16,20,:uni_lag, DynamicSDDiP.Core_Epsilon(perturb=1e-2,integer_relax=false,normalize_direction=false), DynamicSDDiP.MultiCutRegime(), DynamicSDDiP.NoCutSelection(), "C:/Users/cg4102/Documents/julia_logs/CLSP_RR2_Large_O_16_20_Eps.log", 10800, 11111, 12345)
+    #model_starter(7,4,20,:uni_lag, DynamicSDDiP.L₁_Deep(), DynamicSDDiP.MultiCutRegime(), DynamicSDDiP.NoCutSelection(), "C:/Users/cg4102/Documents/julia_logs/Test.log", 10800, 11111, 12345)
+    #model_starter(7,16,20,:uni_lag, DynamicSDDiP.L∞_Deep(), DynamicSDDiP.MultiCutRegime(), DynamicSDDiP.NoCutSelection(), "C:/Users/cg4102/Documents/julia_logs/CLSP_RR_Large_O_16_20_Lsup.log", 10800, 11111, 12345)
+    #model_starter(9,16,20,:uni_lag, DynamicSDDiP.L₁∞_Deep(), DynamicSDDiP.MultiCutRegime(), DynamicSDDiP.NoCutSelection(), "C:/Users/cg4102/Documents/julia_logs/CLSP_RR_Large_O_16_20_L1sup.log", 10800, 11111, 12345)
+    #model_starter(11,16,20,:B, DynamicSDDiP.Core_Midpoint(), DynamicSDDiP.MultiCutRegime(), DynamicSDDiP.NoCutSelection(), "C:/Users/cg4102/Documents/julia_logs/CLSP_RR_Large_O_16_20_B_Multi.log", 10800, 11111, 12345)
+    model_starter(12,4,20,:SB, DynamicSDDiP.Core_Midpoint(), DynamicSDDiP.MultiCutRegime(), DynamicSDDiP.NoCutSelection(), "C:/Users/cg4102/Documents/julia_logs/Test.log", 10800, 11111, 12345)
+    #model_starter(11,16,20,:B, DynamicSDDiP.Core_Midpoint(), DynamicSDDiP.SingleCutRegime(), DynamicSDDiP.NoCutSelection(), "C:/Users/cg4102/Documents/julia_logs/CLSP_RR_Large_O_16_20_B.log", 10800, 11111, 12345)
+    #model_starter(12,16,20,:SB, DynamicSDDiP.Core_Midpoint(), DynamicSDDiP.SingleCutRegime(), DynamicSDDiP.NoCutSelection(), "C:/Users/cg4102/Documents/julia_logs/CLSP_RR_Large_O_16_20.log", 10800, 11111, 12345)
+    #model_starter(12,4,20,:uni_lag, DynamicSDDiP.Core_Midpoint(integer_regime=DynamicSDDiP.NoIntegerRelax(),normalize_direction=true,unbounded_regime=DynamicSDDiP.Unbounded_Opt_SB()), DynamicSDDiP.MultiCutRegime(), DynamicSDDiP.NoCutSelection(), "C:/Users/cg4102/Documents/julia_logs/Test.log", 14400, 11111, 12345)
+    #model_starter(13,4,20,:uni_lag, DynamicSDDiP.Core_Conv(lambda=0.75,copy_regime=DynamicSDDiP.StateSpaceCopy(),normalize_direction=true), DynamicSDDiP.MultiCutRegime(), DynamicSDDiP.NoCutSelection(), "C:/Users/cg4102/Documents/julia_logs/Test.log", 14400, 11111, 12345)
+
+    #model_starter(14,16,20,:uni_lag, DynamicSDDiP.L₁_Deep(), DynamicSDDiP.MultiCutRegime(), DynamicSDDiP.NoCutSelection(), "C:/Users/cg4102/Documents/julia_logs/CLSP_RR_Large_O_16_20_CL_L1.log", 10800, 11111, 12345)
+    #model_starter(15,16,20,:uni_lag, DynamicSDDiP.ChenLuedtke(), DynamicSDDiP.MultiCutRegime(), DynamicSDDiP.NoCutSelection(), "C:/Users/cg4102/Documents/julia_logs/CLSP_RR_Large_O_16_20_CL_Span.log", 10800, 11111, 12345)
+    #model_starter(15,16,20,:uni_lag, DynamicSDDiP.Core_Midpoint(), DynamicSDDiP.MultiCutRegime(), DynamicSDDiP.NoCutSelection(), "C:/Users/cg4102/Documents/julia_logs/CLSP_RR_Large_O_16_20_CL_Mid.log", 10800, 11111, 12345)
+    #model_starter(15,16,20,:uni_lag, DynamicSDDiP.Core_Epsilon(perturb=1e-2,integer_relax=false), DynamicSDDiP.MultiCutRegime(), DynamicSDDiP.NoCutSelection(), "C:/Users/cg4102/Documents/julia_logs/CLSP_RR_Large_O_16_20_CL_Eps.log", 10800, 11111, 12345)
+    #model_starter(15,16,20,:uni_lag, DynamicSDDiP.Core_Relint(), DynamicSDDiP.MultiCutRegime(), DynamicSDDiP.NoCutSelection(), "C:/Users/cg4102/Documents/julia_logs/CLSP_RR_Large_O_16_20_CL_Relint.log", 10800, 11111, 12345)
+    #model_starter(15,16,20,:uni_lag, DynamicSDDiP.L∞_Deep(), DynamicSDDiP.MultiCutRegime(), DynamicSDDiP.NoCutSelection(), "C:/Users/cg4102/Documents/julia_logs/CLSP_RR_Large_O_16_20_CL_Lsup.log", 10800, 11111, 12345)
+
+end
+
+#model_starter_runs()
