+++ examples/CLSP/sample_log_normal.jl
