--- examples/CLSP_larger/starter_no_bin.jl
@@ -1,169 +1,0 @@
-import DynamicSDDiP
-import Infiltrator
-import MathOptInterface
-using Revise
-using Printf
-
-include("algo_config.jl")
-include("scenario_tree.jl")
-include("model.jl")
-include("model_no_bin.jl")
-include("simulation.jl")
-
-function model_starter(
-    num::Int,
-    number_of_stages::Int,
-    number_of_realizations::Int,
-    duality_regime_sym::Symbol,
-    normalization_regime::DynamicSDDiP.AbstractNormalizationRegime,
-    cut_aggregation_regime::DynamicSDDiP.AbstractCutAggregationRegime,
-    cut_selection_regime::DynamicSDDiP.AbstractCutSelectionRegime,
-    log_file::String,
-    time_limit::Int,
-    forward_seed::Int,
-    tree_seed::Int
-)
-
-    try
-        ########################################################################
-        # DEFINE ALGO PARAMS
-        ########################################################################
-        algo_config_output = algo_config(duality_regime_sym, normalization_regime, cut_aggregation_regime, cut_selection_regime, log_file, time_limit, forward_seed)
-        algo_params = algo_config_output.algo_params
-        applied_solvers = algo_config_output.applied_solvers
-
-        ########################################################################
-        # DEFINE MODEL
-        ########################################################################
-        model_output = model_no_bin_set_up(number_of_stages, number_of_realizations, algo_params=algo_params, applied_solvers=applied_solvers, tree_seed=tree_seed)
-        model = model_output.model
-        problem_params = model_output.problem_params
-
-        ########################################################################
-        # SOLVE (TRAIN) MODEL
-        ########################################################################
-        Random.seed!(forward_seed)
-        DynamicSDDiP.solve(model, algo_params, applied_solvers, problem_params)
-
-        ########################################################################
-        # SIMULATE MODEL
-        ########################################################################
-        #simulate(model, algo_params, problem_params, algo_params.simulation_regime)
-
-        ########################################################################
-        # SIMULATE MODEL USING FULL SCENARIO TREE
-        ########################################################################
-        #simulate(model, algo_params, problem_params, DynamicSDDiP.HistoricalSample())
-
-    catch e
-        @printf "Case %d terminated with error" num
-        println()
-        #throw(error(e))
-        showerror(stdout, e, catch_backtrace())
-        println()
-        println("#############################################################")
-        println()
-    end
-
-end
-
-
-function det_equiv_no_bin_starter(
-    num::Int,
-    number_of_stages::Int,
-    number_of_realizations::Int,
-    duality_regime_sym::Symbol,
-    normalization_regime::DynamicSDDiP.AbstractNormalizationRegime,
-    cut_aggregation_regime::DynamicSDDiP.AbstractCutAggregationRegime,
-    cut_selection_regime::DynamicSDDiP.AbstractCutSelectionRegime,
-    log_file::String,
-    time_limit::Int,
-    forward_seed::Int,
-    tree_seed::Int
-)
-
-    try
-        ############################################################################
-        # DEFINE ALGO PARAMS
-        ############################################################################
-        algo_config_output = algo_config(duality_regime_sym, normalization_regime, cut_aggregation_regime, cut_selection_regime, log_file, time_limit, forward_seed)
-        algo_params = algo_config_output.algo_params
-        applied_solvers = algo_config_output.applied_solvers
-
-        ############################################################################
-        # DEFINE MODEL
-        ############################################################################
-        model_output = model_no_bin_set_up(number_of_stages, number_of_realizations, algo_params=algo_params, applied_solvers=applied_solvers, tree_seed=tree_seed)
-        model = model_output.model
-        problem_params = model_output.problem_params
-
-        ############################################################################
-        # SOLVE (TRAIN) MODEL
-        ############################################################################
-        det_equiv = SDDP.deterministic_equivalent(model, Gurobi.Optimizer, time_limit = 7200.0)
-        JuMP.set_objective_sense(det_equiv, MathOptInterface.MIN_SENSE)
-        JuMP.set_optimizer_attribute(det_equiv, "TimeLimit", 7200.0)
-        JuMP.optimize!(det_equiv)
-        print(JuMP.objective_value(det_equiv))
-
-        ############################################################################
-        # LOGGING
-        ############################################################################
-        log_file_handle = open(algo_params.log_file, "a")
-        DynamicSDDiP.print_helper(DynamicSDDiP.print_det_equiv, log_file_handle, problem_params, JuMP.objective_value(det_equiv), JuMP.objective_bound(det_equiv))
-        close(log_file_handle)
-
-    catch e
-        @printf "Case %d (deterministic equivalent) terminated with error" num
-        println()
-        #throw(error(e))
-        showerror(stdout, e, catch_backtrace())
-        println()
-        println("#############################################################")
-        println()
-    end
-
-end
-
-
-function model_starter_runs()
-
-    """
-    Specification of model runs that should be run one after the other (for CLSP larger with 10 state variables).
-    """
-
-    #model_starter(10,16,20,:B, DynamicSDDiP.Core_Midpoint(), DynamicSDDiP.SingleCutRegime(), DynamicSDDiP.NoCutSelection(), "C:/Users/cg4102/Documents/julia_logs/CLSP_Large_B_Multi_2025.log", 10800, 11111, 12345)
-    #model_starter(11,16,20,:SB, DynamicSDDiP.Core_Midpoint(), DynamicSDDiP.SingleCutRegime(), DynamicSDDiP.NoCutSelection(), "C:/Users/cg4102/Documents/julia_logs/CLSP_Large_B_Multi_2025.log", 10800, 11111, 12345)
-    #model_starter(12,16,20,:B, DynamicSDDiP.Core_Midpoint(), DynamicSDDiP.MultiCutRegime(), DynamicSDDiP.NoCutSelection(), "C:/Users/cg4102/Documents/julia_logs/CLSP_Large_B_Multi_2025.log", 10800, 11111, 12345)
-    #model_starter(13,16,20,:SB, DynamicSDDiP.Core_Midpoint(), DynamicSDDiP.MultiCutRegime(), DynamicSDDiP.NoCutSelection(), "C:/Users/cg4102/Documents/julia_logs/CLSP_Large_SB_Multi_2025.log", 10800, 11111, 12345)
-
-    #model_starter(2,16,20,:lag, DynamicSDDiP.Core_Midpoint(), DynamicSDDiP.SingleCutRegime(), DynamicSDDiP.NoCutSelection(), "C:/Users/cg4102/Documents/julia_logs/CLSP_Large_O_16_20_Single_SB_ws03.log", 10800, 11111, 12345)
-    #model_starter(2,16,20,:lag, DynamicSDDiP.Core_Midpoint(), DynamicSDDiP.MultiCutRegime(), DynamicSDDiP.NoCutSelection(), "C:/Users/cg4102/Documents/julia_logs/CLSP_Large_O_16_20_Multi_SB_ws03.log", 10800, 11111, 12345)
-    model_starter(4,16,20,:uni_lag, DynamicSDDiP.Core_Midpoint(copy_regime=DynamicSDDiP.ConvexHullCopy(),unbounded_regime=DynamicSDDiP.Unbounded_Opt_SB()), DynamicSDDiP.MultiCutRegime(), DynamicSDDiP.NoCutSelection(), "C:/Users/cg4102/Documents/julia_logs/CLSP_Large_Mid_2025.log", 10800, 11111, 12345)
-    # model_starter(4,16,20,:uni_lag, DynamicSDDiP.Core_In_Out(integer_relax=false,normalize_direction=false), DynamicSDDiP.MultiCutRegime(), DynamicSDDiP.NoCutSelection(), "C:/Users/cg4102/Documents/julia_logs/CLSP_RR2_Large_O_16_20_InOut.log", 10800, 11111, 12345)
-    model_starter(5,16,20,:uni_lag, DynamicSDDiP.Core_Relint(copy_regime=DynamicSDDiP.ConvexHullCopy(),unbounded_regime=DynamicSDDiP.Unbounded_Opt_SB()), DynamicSDDiP.MultiCutRegime(), DynamicSDDiP.NoCutSelection(), "C:/Users/cg4102/Documents/julia_logs/CLSP_Large_Relint_2025.log", 10800, 11111, 12345)
-    model_starter(16,16,20,:uni_lag, DynamicSDDiP.Core_Epsilon(perturb=1e-2,copy_regime=DynamicSDDiP.ConvexHullCopy(),unbounded_regime=DynamicSDDiP.Unbounded_Opt_SB()), DynamicSDDiP.MultiCutRegime(), DynamicSDDiP.NoCutSelection(), "C:/Users/cg4102/Documents/julia_logs/CLSP_Large_Eps_2025.log", 10800, 11111, 12345)
-    model_starter(7,16,20,:uni_lag, DynamicSDDiP.L₁_Deep(), DynamicSDDiP.MultiCutRegime(), DynamicSDDiP.NoCutSelection(), "C:/Users/cg4102/Documents/julia_logs/CLSP_Large_L1_2025.log", 10800, 11111, 12345)
-    model_starter(7,16,20,:uni_lag, DynamicSDDiP.L∞_Deep(), DynamicSDDiP.MultiCutRegime(), DynamicSDDiP.NoCutSelection(), "C:/Users/cg4102/Documents/julia_logs/CLSP_Large_Lsup_2025.log", 10800, 11111, 12345)
-    #model_starter(9,16,20,:uni_lag, DynamicSDDiP.L₁∞_Deep(), DynamicSDDiP.MultiCutRegime(), DynamicSDDiP.NoCutSelection(), "C:/Users/cg4102/Documents/julia_logs/CLSP_Large_O_16_20_L1sup_SB_ws03.log", 10800, 11111, 12345)
-
-    model_starter(1,16,20,:uni_lag, DynamicSDDiP.Core_Conv(lambda=0.5,copy_regime=DynamicSDDiP.ConvexHullCopy(),unbounded_regime=DynamicSDDiP.Unbounded_Opt_SB()), DynamicSDDiP.MultiCutRegime(), DynamicSDDiP.NoCutSelection(), "C:/Users/cg4102/Documents/julia_logs/CLSP_Large_Conv_50_2025.log", 10800, 11111, 12345)
-    model_starter(1,16,20,:uni_lag, DynamicSDDiP.Core_Conv(lambda=0.75,copy_regime=DynamicSDDiP.ConvexHullCopy(),unbounded_regime=DynamicSDDiP.Unbounded_Opt_SB()), DynamicSDDiP.MultiCutRegime(), DynamicSDDiP.NoCutSelection(), "C:/Users/cg4102/Documents/julia_logs/CLSP_Large_Conv_75_2025.log", 10800, 11111, 12345)
-    model_starter(1,16,20,:uni_lag, DynamicSDDiP.Core_Conv(lambda=0.9,copy_regime=DynamicSDDiP.ConvexHullCopy(),unbounded_regime=DynamicSDDiP.Unbounded_Opt_SB()), DynamicSDDiP.MultiCutRegime(), DynamicSDDiP.NoCutSelection(), "C:/Users/cg4102/Documents/julia_logs/CLSP_Large_Conv_90_2025.log", 10800, 11111, 12345)
-    model_starter(1,16,20,:uni_lag, DynamicSDDiP.Core_Conv(lambda=0.99,copy_regime=DynamicSDDiP.ConvexHullCopy(),unbounded_regime=DynamicSDDiP.Unbounded_Opt_SB()), DynamicSDDiP.MultiCutRegime(), DynamicSDDiP.NoCutSelection(), "C:/Users/cg4102/Documents/julia_logs/CLSP_Large_Conv_99_2025.log", 10800, 11111, 12345)
-    
-    #model_starter(14,16,20,:uni_lag, DynamicSDDiP.L₁_Deep(), DynamicSDDiP.MultiCutRegime(), DynamicSDDiP.CutSelection(), "C:/Users/cg4102/Documents/julia_logs/CLSP_SSC_Large_O_16_20_CL2_L1.log", 10800, 11111, 12345)
-    #model_starter(15,16,20,:uni_lag, DynamicSDDiP.ChenLuedtke(), DynamicSDDiP.MultiCutRegime(), DynamicSDDiP.CutSelection(), "C:/Users/cg4102/Documents/julia_logs/CLSP_SSC_Large_O_16_20_CL2_Span.log", 10800, 11111, 12345)
-    #model_starter(15,16,20,:uni_lag, DynamicSDDiP.Core_Midpoint(integer_relax=false,normalize_direction=false), DynamicSDDiP.MultiCutRegime(), DynamicSDDiP.CutSelection(), "C:/Users/cg4102/Documents/julia_logs/CLSP_RR2_Large_O_16_20_CL2_Mid.log", 10800, 11111, 12345)
-    #model_starter(15,16,20,:uni_lag, DynamicSDDiP.Core_Epsilon(perturb=1e-2,integer_relax=false,normalize_direction=false), DynamicSDDiP.MultiCutRegime(), DynamicSDDiP.CutSelection(), "C:/Users/cg4102/Documents/julia_logs/CLSP_RR2_Large_O_16_20_CL2_Eps.log", 10800, 11111, 12345)
-    #model_starter(15,16,20,:uni_lag, DynamicSDDiP.Core_Relint(normalize_direction=false), DynamicSDDiP.MultiCutRegime(), DynamicSDDiP.CutSelection(), "C:/Users/cg4102/Documents/julia_logs/CLSP_RR2N_Large_O_16_20_CL_Relint.log", 10800, 11111, 12345)
-    #model_starter(15,16,20,:uni_lag, DynamicSDDiP.L∞_Deep(), DynamicSDDiP.MultiCutRegime(), DynamicSDDiP.CutSelection(), "C:/Users/cg4102/Documents/julia_logs/CLSP_RR2_Large_O_16_20_CL2_Lsup.log", 10800, 11111, 12345)
-
-    #model_starter(14,16,20,:uni_lag, DynamicSDDiP.Core_Epsilon(copy_regime=DynamicSDDiP.ConvexHullCopy(),perturb=1e-2,integer_regime=DynamicSDDiP.NoIntegerRelax(),normalize_direction=false,unbounded_regime=DynamicSDDiP.Unbounded_Opt_SB()), DynamicSDDiP.MultiCutRegime(), DynamicSDDiP.NoCutSelection(), "C:/Users/cg4102/Documents/julia_logs/CLSP_Large_O_RR2_16_20_Eps_sbopt_CL.log", 10800, 11111, 12345)
-    #model_starter(14,16,20,:uni_lag, DynamicSDDiP.Core_Relint(copy_regime=DynamicSDDiP.ConvexHullCopy(),normalize_direction=false,unbounded_regime=DynamicSDDiP.Unbounded_Opt_SB()), DynamicSDDiP.MultiCutRegime(), DynamicSDDiP.NoCutSelection(), "C:/Users/cg4102/Documents/julia_logs/CLSP_Large_O_RR2_16_20_Relint_sbopt_CL.log", 10800, 11111, 12345)
-    #model_starter(14,6,20,:uni_lag, DynamicSDDiP.Core_In_Out(integer_relax=false,normalize_direction=false), DynamicSDDiP.MultiCutRegime(), DynamicSDDiP.NoCutSelection(), "C:/Users/cg4102/Documents/julia_logs/CLSP_RR3_4_20_InOut.log", 10800, 11111, 12345)
-    #model_starter(14,16,20,:uni_lag, DynamicSDDiP.Core_Midpoint(copy_regime=DynamicSDDiP.ConvexHullCopy(),integer_regime=DynamicSDDiP.NoIntegerRelax(),normalize_direction=false,unbounded_regime=DynamicSDDiP.Unbounded_Opt_SB()), DynamicSDDiP.MultiCutRegime(), DynamicSDDiP.NoCutSelection(), "C:/Users/cg4102/Documents/julia_logs/CLSP_Large_O_RR2_16_20_Mid_sbopt_CL.log", 10800, 11111, 12345)
-
-end
-
-#model_starter_runs()
