+++ src/DynamicSDDiP.jl
@@ -56,9 +56,4 @@ include("backwardPass_classic.jl")
 
 include("lagrange_unified.jl")
 
-include("backwardPass_aggregated.jl")
-include("duals_aggregated.jl")
-include("lagrange_unified_preparation_aggregated.jl")
-include("lagrange_unified_aggregated.jl")
-
 end
